+++ Backend/Gateway/include/interfaces/IVerifier.h
@@ -6,7 +6,7 @@
 
 class IVerifier {
   public:
-    virtual std::optional<int> verifyTokenAndGetUserId(const std::string& token) = 0;
+    virtual std::optional<long long> verifyTokenAndGetUserId(const std::string& token) = 0;
     virtual ~IVerifier() = default;
 };
